+++ packages/abi/lib.esm/index.js
@@ -1,5 +1,5 @@
 "use strict";
 import { ConstructorFragment, EventFragment, FormatTypes, Fragment, FunctionFragment, ParamType } from "./fragments";
 import { AbiCoder, defaultAbiCoder } from "./abi-coder";
-import { Indexed, Interface } from "./interface";
-export { ConstructorFragment, EventFragment, Fragment, FunctionFragment, ParamType, FormatTypes, AbiCoder, defaultAbiCoder, Interface, Indexed };
+import { Indexed, Interface, LogDescription, TransactionDescription } from "./interface";
+export { ConstructorFragment, EventFragment, Fragment, FunctionFragment, ParamType, FormatTypes, AbiCoder, defaultAbiCoder, Interface, Indexed, LogDescription, TransactionDescription };