+++ packages/base64/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "base64/5.0.0-beta.131";
+export declare const version = "base64/5.0.0-beta.132";