+++ packages/hardware-wallets/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "hardware-wallets/5.0.0-beta.3";
+export declare const version = "hardware-wallets/5.0.0-beta.4";