+++ packages/random/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "random/5.0.0-beta.133";
+export declare const version = "random/5.0.0-beta.134";