+++ packages/testcases/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "testcases/5.0.0-beta.138";
+export declare const version = "testcases/5.0.0-beta.139";